Fix height issue
The height of a widget may depend on the layout. But the layout is
affected by the previously rendered widgets. For example, the `Select`
widget adds 1 to the height if the line_offset is more than 1, since it
needs to go to the next line, but if it is given a layout that has
line_offset 0, during height(), and with line_offset > 0 during render,
this may cause issues when rendering at the end of the screen.

So now height() requires layout to be edited similar to render()