Make TestBackend render everything instead of asserting correct ops
If the same characters are written through 2 writes instead of 1, the
previous version would error. This version however would not, which
makes more sense since what the user sees is the same.

Tests also now use `insta` for snapshots of the view which is more
manageable then having hard-coded values there.