Fix #2
`Widget::cursor_pos` used to return the position relative to the root
widget component. However, components such as `Input` would adjust the
layout to include the `Prompt` and then calculate `cursor_pos`. Since
the `cursor_pos` returned was relative to the end of the `Prompt`, if
the `Prompt` was more than one line, it would calculate the wrong
location.

This commit now makes `cursor_pos` return the position relative to the
screen.