Change rendering of the default value in input, int and float
Earlier, the default value would just be shown on the side at all times.
This is even if the default will not be selected which can be
misleading. This change shows the default as greyed out text in the
input itself. It also allows pressing 'Tab' to make the current input
the default if the current input value is the start of the default.

Implements #4