Show a message when the prompt is too big for the screen
Earlier, when the prompt was the large, the program would panic and
crash. Now, it displays a message instead.

The ideal solution would be to be able to render things off-screen as
well, however this has a lot of scope for bugs and adds a lot of
complexity to the backends. As such, unless there are any other future
issues regarding this problem, this solution seems like a good enough
tradeoff.

Fixes #17